requirements.txt
click
discord.py
sqlalchemy
dpytest
pytest
